+++ HISTORY.md
@@ -1,5 +1,13 @@
 # History
 
+# not yet published, version 10.0.2
+
+- Fixed #2156: simplify expressions like `-1 / (-x)` to `1/x`. Thanks @ony3000.
+- Fixed #2363: remove a redundant part of the regex to split a number.
+- Fixed #2291: add support for fractions in function `intersect`. 
+  Thanks @thetazero.
+
+
 # 2021-12-22, version 10.0.1
 
 - Fix #1681: function `gamma` giving inaccurate complex results in some cases.